moving some classes to io.chymyst.util
